# Override TextWriter.Write{Line}{Async} on StringWriter