closes https://github.com/dotnet/corefx/issues/30502

/cc @stephentoub 

Before

 System.Runtime.Extensions.Performance.Tests.dll  | Metric   | Unit | Iterations | Average | STDEV.S |     Min |     Max
:------------------------------------------------ |:-------- |:----:|:----------:| -------:| -------:| -------:| -------:
 System.IO.Tests.Perf_StringWriter.Write          | Duration | msec |     97     | 103.497 |  11.100 |  68.060 | 138.213
 System.IO.Tests.Perf_StringWriter.WriteAsync     | Duration | msec |    100     | 100.057 |   9.167 |  77.288 | 119.253
 System.IO.Tests.Perf_StringWriter.WriteLine      | Duration | msec |     85     | 118.157 |   5.250 |  98.278 | 125.735
 System.IO.Tests.Perf_StringWriter.WriteLineAsync | Duration | msec |     66     | 152.739 |   5.793 | 139.163 | 184.234

After 

 System.Runtime.Extensions.Performance.Tests.dll  | Metric   | Unit | Iterations | Average | STDEV.S |    Min |    Max
:------------------------------------------------ |:-------- |:----:|:----------:| -------:| -------:| ------:| ------:
 System.IO.Tests.Perf_StringWriter.Write          | Duration | msec |    100     |  59.515 |   9.363 | 38.556 | 78.121
 System.IO.Tests.Perf_StringWriter.WriteAsync     | Duration | msec |    100     |  63.386 |   8.228 | 42.660 | 75.798
 System.IO.Tests.Perf_StringWriter.WriteLine      | Duration | msec |    100     |  58.092 |   8.738 | 38.085 | 76.327
 System.IO.Tests.Perf_StringWriter.WriteLineAsync | Duration | msec |    100     |  63.826 |   8.457 | 43.962 | 76.391