@MarcoRossignoli 2018-06-28
@stephentoub i see that CI now it's ok...is for this reason https://github.com/dotnet/corefx/pull/30447#issuecomment-397836685?
Is there a way to understand it by myself(for future,  i've been thinking it's my fault, some docs, issues, PR around)?