@stephentoub 2018-06-28
> is for this reason

Yes, the Corelib changes hadn't yet propagated to the .NET Native bits.